+++ tutorials/ngsi_v2/e3_context_entities/e3_context_entities.py
@@ -57,7 +57,7 @@ SERVICE = 'filip_tutorial'
 SERVICE_PATH = '/<your_path>'
 
 # ToDo: Path to json-files to store entity data for follow up exercises,
-#  e.g. ./e3_my_entities.json
+#  e.g. ../e3_my_entities.json
 WRITE_ENTITIES_FILEPATH = Path("")
 
 # ## Main script